requirements.txt
streamlit
dnspython
beautifulsoup4
requests
urllib3
